crm: factor out giving up the watchdog protection
See the added comment for full details why the watchdog protection of
the CRM needs less strict safety requirements compared to the one of
an (active) LRM.

In short, CRM does not manages services but directs them through the
manager_status state-file. This means the watchdog mainly protects it
from a hung system where locks would timeout before writing the state
out and thus a race with the new CRM would happen. So the CRM
basically can always give up the watchdog safely when it stops being
the active CRM anyway.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>