fix #5243: make CRM go idle after ~15 min of no service being configured
This is mostly for convenience for when one does a quick HA evaluation
and then removes all services again, the biggest visible effect is
that there will be no status updates once the CRM is idle, reducing
some mid-frequent updates to pmxcfs.

As the CRM never lets the watchdog run out pro-actively to trigger
fencing, this won't have much of a difference w.r.t. "accidental"
self-fencing on the common outage situation that happen in practice,
i.e. quorum loss due to network outage or corosync getting
misconfigured, that's also why this was not considered when adding the
auto-idling for LRM back in commit 2105170 ("LRM: release lock and
close watchdog if no service configured for >10min").
In short, the watchdog for CRM is mostly here to avoid a situation
where the process of the currently active CRM hangs, or does not get
scheduled for a while such that another CRM becomes active only that
the previous one then resumes and still thinks it is the active one
and, e.g., writes out a outdated manager_status file; there are some
other situations, but it's always similar reason.

Compared to the LRM idle mechanism we require more rounds for the CRM
to go idle (90 for CRM vs 60 for LRM), the reason here is that the LRM
needs an active CRM for some operations to progress in the FSM, so
just waiting a bit longer for the CRM is enough to ensure that this
can happen.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>