env: add any_pending_crm_command method
Add helper that returns if the CRM command queue holds any commands
without altering the state of the queue at all, unlike the existing
read_crm_commands method would do.

This will be used to check if a CRM needs to become active when there
are pending CRM commands but no master seems to process them as all
are idle/offline.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>