crm: get active if there are nodes that need to leave maintenance
This is mostly cosmetic, because as long as there are configured
services a CRM would get active anyway. But it can happen that a
maintenance mode is left-over and all services got removed a fresh
cluster start then will keep all CRMs as idle and thus never clear the
maintenance state.

This can be especially confusing now, as a recent pve-manager commit
993d05abc ("api/ui: include the node ha status in resources call and
show as icon") started to show the maintenance mode as icon in the web
UI, thus making this blip much more prominent.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>