LRM: count incoming migrations towards a nodes active resources
This ensures an idle LRM starts up as soon as possible on an incoming
migration and makes it more consistent with the CRM behavior, which
also counts migrating resources towards the usage for both source and
target nodes.

This greatly improves the behavior for long running migrations and
node (quorum) failures, as there the watchdog gets already active once
the migration starts, not only once it finished.
The effects of this change can be verified and observed nicely from
the changes in our excellent regression and behavior test suite.

This could be made even more strict by adding a wait barrier to ensure
the target LRM is fully active before actually starting the migration,
but that also would delay such migrations even longer, and users would
like to have lower latencies on LRM command executions now already.
Even if we decide to adopt such a barrier (on either source or
target), the patch here would be required as prerequisite in any way,
so improve the status quo now already.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>