watchdog: switch to epoll_create1 and tell it to close FD on exit
The epoll_create(size) is semi-deprecated, it's size argument is
ignored, it just needs to be positive for backward compat, besides
that the kernel will auto (re-)size this internally.

The newer epoll_create1 dropped that argument and added a flags one,
so switch over to that function and set close on exec to avoid leaking
an FD to e.g. the journal-sync fork+exec invocations. This is safe to
do for epoll as it will only deconstruct once all FD (references) are
closed.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>