watchdog-mux: Restore if guard for watchdog updates
In commit c26b474 the if condition was removed in favor of the if guard
on the higher level. However, this introduced functional change, the
watchdog now updates immediately after update_watchdog is set to 0,
potentially causing fencing after x+1 seconds instead of x seconds. This
behavior is incorrect because updating the watchdog after declaring no
more updates makes no logical sense.

Signed-off-by: Maximiliano Sandoval <m.sandoval@proxmox.com>
Reported-by: Friedrich Weber <f.weber@proxmox.com>
Link: https://lore.proxmox.com/20250723105518.272549-1-m.sandoval@proxmox.com