manager: apply node affinity rules when selecting service nodes
Replace the HA group mechanism with the functionally equivalent node
affinity rules' get_node_affinity(...), which enforces the node affinity
rules defined in the rules config.

This allows the $groups parameter to be replaced with the $rules
parameter in select_service_node(...) as all behavior of the HA groups
is now encoded in $service_conf and $rules.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20250730175957.386674-12-d.kral@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>