cli: expose ha rules api endpoints to ha-manager cli
Expose the HA rules API endpoints through the CLI in its own subcommand.

The names of the subsubcommands are chosen to be consistent with the
other commands provided by the ha-manager CLI for HA resources and
groups, but grouped into a subcommand.

The properties specified for the 'rules config' command are chosen to
reflect the columns from the WebGUI for the HA rules.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20250730175957.386674-15-d.kral@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>