sim: do not create default groups config
As none of the existing HA test cases rely on the default HA groups and
the pve-ha-simulator does not have any GUI integration of HA groups
anyway, remove them from being created.

This is done, because in an upcoming patch, which persistently migrates
HA groups to node affinity rules, it would unnecessarily fire the
migration for every default group config.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20250730175957.386674-16-d.kral@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>