manager: handle resource affinity rules in manual migrations
Make any manual user migration of a resource follow the resource
affinity rules it is part of. That is:

- prevent a resource to be manually migrated to a node, which contains a
  resource, that the resource must be kept separate from (negative
  resource affinity).

- make resources, which must be kept together (positive resource
  affinity), migrate to the same target node, and

The log information here is only redirected to the HA Manager node's
syslog, so user-facing endpoints need to implement this logic as well to
give users adequate feedback about these actions.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20250730181428.392906-9-d.kral@proxmox.com