manager: support older versioning schema for ha group migration
Prior to the Proxmox VE 8.x series, the pve-manager is versioned in the
format of `M.m-p`. Versions prior to Proxmox VE 8.x are already EOL'd,
but this change prevents any "Use of uninitialized value" log messages
if one of the nodes is on a version with such a versioning scheme.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20250804142212.212744-1-d.kral@proxmox.com