manager: fix precedence in mixed resource affinity rules usage
Strict positive resource affinity rules narrow down the possible nodes
to a single candidate for a HA resource A, which is the node, where the
most of the HA resources in the positive affinity rule are already
running on and in case of a tie the alphabetically first node is chosen.

If the chosen node contains a HA resource B, which is in negative
affinity with the HA resource A, then $pri_nodes will become empty and
will not result in any migration.

Therefore, apply the negative resource affinity rules before the
positive resource affinity rules to prevent the premature pruning of
nodes.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20250919140856.1361124-3-d.kral@proxmox.com