implement static service stats cache
As the HA Manager builds the static load scheduler, it queries the
services' static usage by reading and parsing the static guest configs
individually, which can take significantly more time with respect to how
many HA resources are in an actively managed state.

PVE::Cluster exposes an efficient interface to gather a set of
properties from one or all guest configs [0]. This is used here to build
a rather short-lived cache on every (re)initialization of the static
load scheduler to avoid parsing guest configs individually.

[0] pve-cluster cf1b19d (add get_guest_config_property IPCC method)

Suggested-by: Fiona Ebner <f.ebner@proxmox.com>
Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Reviewed-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/20251114100641.92919-13-d.kral@proxmox.com