compile ha rules to a more compact representation
The initial implementation of fetching a ha resource's rules in the
scheduler and other users is rather inefficient as it needs to go
through the raw rules hash map and gathers quite a lot of information
which doesn't change unless the HA rules need to be re-translated
anyway.

Therefore add a compile stage when re-translating the HA rules, which
compiles each rule plugin's rules into a more compact data structure.
This reduces the unnecessary recomputation of invariant data done in the
hot path of the scheduler's select_service_node(...).

A benchmark done with NYTProf [0] shows the following runtime
improvement, where each duration is the average time needed per call.
The benchmark was done by starting up a 3-node cluster with 1200
configured HA resources, where each HA resource is in a node affinity
rule and a negative resource affinity to two other HA resources.

+----------------------------------+--------+-------+
| Function                         | Before | After |
+----------------------------------+--------+-------+
| select_service_node              | 10.5ms | 44µs  |
| get_node_affinity                | 5.26ms | 10µs  |
| get_resource_affinity            | 5.17ms | 12µs  |
+----------------------------------+--------+-------+

[0] https://metacpan.org/pod/Devel::NYTProf

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20251114111828.160294-11-d.kral@proxmox.com