sim: hardware: remove service default groups
These groups are not enforced unless these are explicitly defined in the
HA group config, but even then it would be rather unexpected for users
if they haven't explicitly provided that information in the first place.

Furthermore these are neither used in test cases nor displayed or
modifiable in the pve-ha-simulator, so remove them.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20251117172147.533969-2-d.kral@proxmox.com